==> notebooks/02_count_folders.ipynb <==
import os

# Percorso della cartella da analizzare
directory_path = "../data/raw/train/openpose_output_train/json"

try:
    # Verifica che il percorso esista e sia una cartella
    if os.path.isdir(directory_path):
        # Elenca tutti gli elementi nella cartella e conta solo le sottocartelle
        folder_count = sum(
            1
            for item in os.listdir(directory_path)
            if os.path.isdir(os.path.join(directory_path, item))
        )
        print(f"Numero di cartelle in '{directory_path}': {folder_count}")
    else:
        print(
            f"Il percorso specificato non è una cartella o non esiste: '{directory_path}'"
        )
except FileNotFoundError:
    print(f"Errore: Percorso non trovato '{directory_path}'")
except Exception as e:
    print(f"Si è verificato un errore: {e}")

import os

# Percorso della cartella da analizzare
directory_path = "../data/raw/train/openpose_output_train/video"

try:
    # Verifica che il percorso esista e sia una cartella
    if os.path.isdir(directory_path):
        # Elenca tutti gli elementi nella cartella e conta solo i file
        file_count = sum(
            1
            for item in os.listdir(directory_path)
            if os.path.isfile(os.path.join(directory_path, item))
        )
        print(f"Numero di file in '{directory_path}': {file_count}")
    else:
        print(
            f"Il percorso specificato non è una cartella o non esiste: '{directory_path}'"
        )
except FileNotFoundError:
    print(f"Errore: Percorso non trovato '{directory_path}'")
except Exception as e:
    print(f"Si è verificato un errore: {e}")
